#943 FIX upper and lower case issues in the seed name
